backend/requirements.txt
distlib==0.3.7
gqlalchemy==1.4.1
fastapi==0.95.0
networkx==2.8.4
pandas==2.0.3
uvicorn==0.22.0
python-dotenv==1.0.0
